ldataクラスを削除　各メソッドを整理
