+++ 05.ls/lib/output.rb
@@ -2,46 +2,32 @@
 
 require_relative './filename'
 require_relative './command'
-require_relative './ldata'
 
 class Output
+  attr_reader :command
+
   COLUMN = 3
   FILE_TYPE = { 'file' => '-', 'directory' => 'd', 'link' => 'l' }.freeze
   PERMISSION = { '0' => '---', '1' => '--x', '2' => '-w-', '3' => '-wx', '4' => 'r--', '5' => 'r-x', '6' => 'rw-',
                  '7' => 'rwx' }.freeze
 
   def initialize
-    filename = FileName.new
-    @array_filenames = filename.make_array
-    @ls_files = filename.make_lsfile
-    @command = filename.command
+    @filename = FileName.new
   end
 
   def output_normal
-    @array_filenames << nil until (@array_filenames.size % COLUMN).zero?
-    sliced_array = @array_filenames.each_slice(@array_filenames.size / COLUMN).to_a
-    transposed_array = sliced_array.transpose
-    transposed_array.each do |names|
-      names.each do |name|
-        print name.to_s.ljust(30)
-      end
-      print "\n"
-    end
-  end
-
-  def output_total
-    total = @array_filenames.sum {|arr| File.stat(arr).blocks }
-    puts "total #{total}"
+    @filename.transposed_array
   end
 
   def output_l
-    output_total
-    ldata = Ldata.new(@ls_files)
-    ldata.make_data_l.each { |data| puts data }
+    @filename.output_total
+    @ls_files = @filename.make_lsfile
+    l_list = @ls_files.map(&:l_line)
+    l_list.each { |data| puts data }
   end
 
   def output
-    if @command.command_l?
+    if @filename.command.command_l?
       output_l
     else
       output_normal
